// src/app/country/interfaces/country.interface.ts
export enum Region {
  Africa    = 'Africa',
  Americas  = 'Americas',
  Asia      = 'Asia',
  Europe    = 'Europe',
  Oceania   = 'Oceania'
}

export interface CountrySmall {
  name: string;
  cca3: string;
  borders?: string[];
}

export interface Name {
  common:     string;
}

// Generated by https://quicktype.io

export interface Country {
  name:         Name;
  tld:          string[];
  cca2:         string;
  ccn3:         string;
  cca3:         string;
  cioc?:        string;
  independent:  boolean;
  status:       Status;
  unMember:     boolean;
  currencies?:  Currencies;
  idd:          Idd;
  capital?:     string[];
  altSpellings: string[];
  region:       Region;
  subregion?:   string;
  languages?:   { [key: string]: string };
  translations: { [key: string]: Translation };
  latlng:       number[];
  landlocked:   boolean;
  borders?:     string[];
  area:         number;
  demonyms?:    Demonyms;
  flag:         string;
  maps:         Maps;
  population:   number;
  fifa?:        string;
  car:          Car;
  timezones:    string[];
  continents:   Continent[];
  flags:        Flags;
  coatOfArms:   CoatOfArms;
  startOfWeek:  StartOfWeek;
  capitalInfo:  CapitalInfo;
  gini?:        { [key: string]: number };
  postalCode?:  PostalCode;
}

export interface CapitalInfo {
  latlng?: number[];
}

export interface Car {
  signs?: string[];
  side:   Side;
}

export enum Side {
  Left = "left",
  Right = "right",
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export enum Continent {
  Africa = "Africa",
  Antarctica = "Antarctica",
  Asia = "Asia",
  NorthAmerica = "North America",
  Oceania = "Oceania",
  SouthAmerica = "South America",
}

export interface Currencies {
  AFN?: TartuGecko;
  SDG?: Sdg;
  EUR?: TartuGecko;
  USD?: TartuGecko;
  NZD?: TartuGecko;
  TZS?: TartuGecko;
  MOP?: TartuGecko;
  XOF?: TartuGecko;
  UGX?: TartuGecko;
  TTD?: TartuGecko;
  DZD?: TartuGecko;
  MRU?: TartuGecko;
  GEL?: TartuGecko;
  ANG?: TartuGecko;
  NAD?: TartuGecko;
  ZAR?: TartuGecko;
  MUR?: TartuGecko;
  AUD?: TartuGecko;
  KID?: TartuGecko;
  XAF?: TartuGecko;
  QAR?: TartuGecko;
  KYD?: TartuGecko;
  BWP?: TartuGecko;
  EGP?: TartuGecko;
  ILS?: TartuGecko;
  JOD?: TartuGecko;
  VND?: TartuGecko;
  GYD?: TartuGecko;
  TVD?: TartuGecko;
  PGK?: TartuGecko;
  BSD?: TartuGecko;
  CVE?: TartuGecko;
  OMR?: TartuGecko;
  VES?: TartuGecko;
  BOB?: TartuGecko;
  CAD?: TartuGecko;
  KGS?: TartuGecko;
  AMD?: TartuGecko;
  CDF?: TartuGecko;
  SZL?: TartuGecko;
  NGN?: TartuGecko;
  MWK?: TartuGecko;
  STN?: TartuGecko;
  FKP?: TartuGecko;
  GTQ?: TartuGecko;
  BBD?: TartuGecko;
  GHS?: TartuGecko;
  NIO?: TartuGecko;
  AOA?: TartuGecko;
  DOP?: TartuGecko;
  LYD?: TartuGecko;
  KWD?: TartuGecko;
  BHD?: TartuGecko;
  TMT?: TartuGecko;
  LRD?: TartuGecko;
  PHP?: TartuGecko;
  XCD?: TartuGecko;
  INR?: TartuGecko;
  MMK?: TartuGecko;
  BZD?: TartuGecko;
  MNT?: TartuGecko;
  SAR?: TartuGecko;
  ARS?: TartuGecko;
  XPF?: TartuGecko;
  YER?: TartuGecko;
  GBP?: TartuGecko;
  SHP?: TartuGecko;
  BRL?: TartuGecko;
  SSP?: TartuGecko;
  THB?: TartuGecko;
  TRY?: TartuGecko;
  BMD?: TartuGecko;
  BDT?: TartuGecko;
  SGD?: TartuGecko;
  KES?: TartuGecko;
  UYU?: TartuGecko;
  DKK?: TartuGecko;
  LBP?: TartuGecko;
  BTN?: TartuGecko;
  MYR?: TartuGecko;
  KZT?: TartuGecko;
  TND?: TartuGecko;
  GMD?: TartuGecko;
  SYP?: TartuGecko;
  GNF?: TartuGecko;
  MZN?: TartuGecko;
  WST?: TartuGecko;
  TWD?: TartuGecko;
  KPW?: TartuGecko;
  DJF?: TartuGecko;
  SRD?: TartuGecko;
  RWF?: TartuGecko;
  KRW?: TartuGecko;
  JMD?: TartuGecko;
  MVR?: TartuGecko;
  PYG?: TartuGecko;
  IDR?: TartuGecko;
  MGA?: TartuGecko;
  IQD?: TartuGecko;
  HNL?: TartuGecko;
  MAD?: TartuGecko;
  HKD?: TartuGecko;
  LKR?: TartuGecko;
  HTG?: TartuGecko;
  CUC?: TartuGecko;
  CUP?: TartuGecko;
  TJS?: TartuGecko;
  CKD?: TartuGecko;
  COP?: TartuGecko;
  UZS?: TartuGecko;
  PEN?: TartuGecko;
  PKR?: TartuGecko;
  CLP?: TartuGecko;
  ZMW?: TartuGecko;
  SCR?: TartuGecko;
  CRC?: TartuGecko;
  KMF?: TartuGecko;
  JPY?: TartuGecko;
  FJD?: TartuGecko;
  MXN?: TartuGecko;
  KHR?: TartuGecko;
  IRR?: TartuGecko;
  AED?: TartuGecko;
  AZN?: TartuGecko;
  VUV?: TartuGecko;
  ERN?: TartuGecko;
  NPR?: TartuGecko;
  PAB?: TartuGecko;
  SBD?: TartuGecko;
  ZWL?: TartuGecko;
  AWG?: TartuGecko;
  BND?: TartuGecko;
  LAK?: TartuGecko;
  TOP?: TartuGecko;
  LSL?: TartuGecko;
  SOS?: TartuGecko;
  SLL?: TartuGecko;
  BIF?: TartuGecko;
  ETB?: TartuGecko;
  CNY?: TartuGecko;
}

export interface TartuGecko {
  name:   string;
  symbol: string;
}

export interface Sdg {
  name: string;
}

export interface Demonyms {
  eng:  Eng;
  fra?: Eng;
}

export interface Eng {
  f: string;
  m: string;
}

export interface Flags {
  png:  string;
  svg:  string;
  alt?: string;
}

export interface Idd {
  root?:     string;
  suffixes?: string[];
}

export interface Maps {
  googleMaps:     string;
  openStreetMaps: string;
}

export interface Name {
  common:      string;
  official:    string;
  nativeName?: { [key: string]: Translation };
}

export interface Translation {
  official: string;
  common:   string;
}

export interface PostalCode {
  format: string;
  regex?: string;
}

export enum StartOfWeek {
  Monday = "monday",
  Saturday = "saturday",
  Sunday = "sunday",
}

export enum Status {
  OfficiallyAssigned = "officially-assigned",
}